+++ packages/nuxt-vitest/src/config.ts
@@ -1,6 +1,8 @@
-import type { Nuxt } from '@nuxt/schema'
+import type { Nuxt, ViteConfig } from '@nuxt/schema'
 import type { InlineConfig as VitestConfig } from 'vitest'
 import { InlineConfig, mergeConfig, defineConfig } from 'vite'
+import vuePlugin from '@vitejs/plugin-vue'
+import viteJsxPlugin from '@vitejs/plugin-vue-jsx'
 
 interface GetVitestConfigOptions {
   nuxt: Nuxt
@@ -46,11 +48,29 @@ async function startNuxtAndGetViteConfig(rootDir = process.cwd()) {
   return promise
 }
 
+const vuePlugins = {
+  'vite:vue': [vuePlugin, 'vue'],
+  'vite:vue-jsx': [viteJsxPlugin, 'vueJsx'],
+} as const
+
 export async function getVitestConfigFromNuxt(
   options?: GetVitestConfigOptions
 ): Promise<InlineConfig & { test: VitestConfig }> {
   if (!options) options = await startNuxtAndGetViteConfig()
-  options.viteConfig.plugins = options.viteConfig.plugins?.filter(p => (p as any)?.name !== 'nuxt:import-protection')
+  options.viteConfig.plugins = options.viteConfig.plugins || []
+  options.viteConfig.plugins = options.viteConfig.plugins.filter(
+    p => (p as any)?.name !== 'nuxt:import-protection'
+  )
+
+  for (const name in vuePlugins) {
+    if (!options.viteConfig.plugins?.some(p => (p as any)?.name === name)) {
+      const [plugin, key] = vuePlugins[name as keyof typeof vuePlugins]
+      options.viteConfig.plugins.unshift(
+        // @ts-expect-error mismatching component options
+        plugin((options.viteConfig as ViteConfig)[key])
+      )
+    }
+  }
 
   return {
     ...options.viteConfig,
@@ -58,6 +78,22 @@ export async function getVitestConfigFromNuxt(
       ...options.viteConfig.server,
       middlewareMode: false,
     },
+    plugins: [
+      ...options.viteConfig.plugins,
+      // TODO: https://github.com/nuxt/nuxt/pull/20639
+      {
+        name: 'disable-auto-execute',
+        enforce: 'pre',
+        transform(code, id) {
+          if (id.match(/nuxt3?\/.*\/entry\./)) {
+            return code.replace(
+              /(?<!vueAppPromise = )entry\(\)\.catch/,
+              'Promise.resolve().catch'
+            )
+          }
+        },
+      },
+    ],
     test: {
       ...options.viteConfig.test,
       dir: options.nuxt.options.rootDir,