+++ packages/vitest-environment-nuxt/src/index.ts
@@ -24,7 +24,7 @@ export default <Environment>{
       config: {
         public: {},
         app: { baseURL: '/' },
-        ...environmentOptions?.nuxtRuntimeConfig
+        ...environmentOptions?.nuxtRuntimeConfig,
       },
       data: {},
       state: {},
@@ -39,6 +39,8 @@ export default <Environment>{
       win.IntersectionObserver ||
       class IntersectionObserver {
         observe() {}
+        unobserve() {}
+        disconnect() {}
       }
 
     const h3App = createApp()
@@ -66,7 +68,7 @@ export default <Environment>{
     })
 
     // @ts-ignore
-    await import('#app/entry')
+    await import('#app/entry').then(r => r.default())
 
     return {
       // called after all tests with this env have been run