+++ playground/app.vue
@@ -1,9 +1,15 @@
 <template>
   <LazySomeComponent />
-  <!-- TODO: <NuxtPage /> -->
-  <NuxtLink to="/test">Test link</NuxtLink>
+  <GlobalComponent />
+  <NuxtPage />
+  <NuxtLink to="/test">
+    Test link
+  </NuxtLink>
 </template>
 
 <script setup>
 await new Promise(resolve => setTimeout(resolve, 1))
+useHead({
+  title: 'My App'
+})
 </script>