+++ playground/tests/nuxt/mock-component-2.spec.ts
@@ -8,7 +8,8 @@ it('should mock', async () => {
   const component = await mountSuspended(App)
   expect(component.html()).toMatchInlineSnapshot(`
     "<div> Mocked 1 * 2 = 2</div>
-    <!-- TODO: <NuxtPage /> -->
-    <a href=\\"/test\\">Test link</a>"
+    <div> I am a global component </div>
+    <div>Index page</div>
+    <a href=\\"/test\\"> Test link </a>"
   `)
 })